+++ src/Main.java
@@ -1,8 +1,8 @@
 import person.Person;
-import person.PersonGrandeurComparator;
 
 import java.util.ArrayList;
 import java.util.Collections;
+import java.util.Comparator;
 import java.util.List;
 
 public class Main {
@@ -12,12 +12,28 @@ public class Main {
         noblePeople.add(new Person("Софья", "Августа Фредерика Анхальт-Цербстская", 67));
         noblePeople.add(new Person("Максимилиана", "Вильгельмина Августа София Мария Гессен-Прирейнская", 55));
         noblePeople.add(new Person("Виктория", "Алиса Елена Луиза Беатриса Гессен-Дармштадская", 46));
-        noblePeople.add (new Person("Фридерика", "Луиза Шарлотта Вильгельмина Прусская", 62));
+        noblePeople.add(new Person("Фридерика", "Луиза Шарлотта Вильгельмина Прусская", 62));
 
-        Collections.sort(noblePeople, new PersonGrandeurComparator(5)); // сортируем список, значимое количество - 5
-        System.out.println(noblePeople); // все работает корректно
+        Comparator<Person> comparator = new Comparator<Person>() {
+            @Override
+            public int compare(Person o1, Person o2) {
+                int o1SurnameLength = o1.getSurname().split("-| ").length;
+                int o1SignificantLength = Math.min(o1SurnameLength, 5);
 
-        Collections.sort(noblePeople, new PersonGrandeurComparator(3)); // сортируем список, значимое количество - 3
+                int o2SurnameLength = o2.getSurname().split("-| ").length;
+                int o2SignificantLength = Math.min(o2SurnameLength, 5);
+
+                if (o1SignificantLength < o2SignificantLength) {
+                    return 1;
+                } else if (o1SignificantLength > o2SignificantLength) {
+                    return -1;
+                } else {
+                    return Integer.compare(o2.getAge(), o1.getAge());
+                }
+            }
+        };
+
+        Collections.sort(noblePeople, comparator); // сортируем список, значимое количество - 5
         System.out.println(noblePeople); // все работает корректно
     }
 }
